fixed bag with points in spb

+++ app/src/main/java/com/example/geochallenge/data/GeochallengeService.kt
@@ -54,6 +54,6 @@ interface GeochallengeService {
     fun postStats(
         mapId: Int,
         taskName: String,
-        distance: Int
+        distance: Double
     ): Completable
 }
