+++ app/src/main/java/com/example/geochallenge/game/controlers/MultiplayerGameControler.kt
@@ -37,7 +37,7 @@ class MultiplayerGameControler @Inject constructor() : GameControler {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun postGameStats(taskName: String, distance: Int): Completable {
+    override fun postGameStats(taskName: String, distance: Double): Completable {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
